feat: Add ExampleFoodListView and JSONLoader
The commit adds the `ExampleFoodListView` struct and `JSONLoader` class to the project. The `ExampleFoodListView` is a SwiftUI view that displays a list of foods fetched from JSON data using the `JSONLoader` class. This commit also includes the addition of the `FoodTags.json` file, which contains the JSON data for food tags.