Setup data: JSON reader and Core Data (#1)
* feat: Add ExampleFoodListView and JSONLoader

The commit adds the `ExampleFoodListView` struct and `JSONLoader` class to the project. The `ExampleFoodListView` is a SwiftUI view that displays a list of foods fetched from JSON data using the `JSONLoader` class. This commit also includes the addition of the `FoodTags.json` file, which contains the JSON data for food tags.

* feat: Add Core Data models for User, WeightLog, and FoodIntake

This commit adds the Core Data models for User, WeightLog, and FoodIntake. Each model is represented by a separate Swift file and includes the necessary properties and relationships. These models will be used to store and manage user data related to weight logs and food intake. Added some examples.